PTSD-1270 Added "sharedMemory" report to the SBC_STATUS_REPORT to help monitor memory usage while running.  We may want to expand the new status report to pa-ws to support more metrics like this, as it is useful to monitor these key runtime metrics, not just log them.

+++ src/basecaller/scripts/start_pa_mongo.sh
@@ -56,7 +56,7 @@ then
        exit 1
     fi
     trc_output="--outputtrcfile $TRACE_OUTPUT"
-    rm -f $TRACE_OUTPUT
+    if [[ ! $ECHO ]] ; then rm -f $TRACE_OUTPUT ; fi
 fi
 if [[ $BAZ_OUTPUT != "" ]]
 then
@@ -101,6 +101,7 @@ cat <<HERE > $tmpjson
          "primerScaler": ${PRIMERSCALER}
     }
   }
+  // , "system":{"analyzerHardware":"Host"}
 } 
 HERE
